Fibonacci sequence generator
As a junior developer, you are required to create a Fibonacci sequence generator.
• The Fibonacci sequence is a series of numbers where a number is found by adding up the two
numbers before it. Starting with 0 and 1, the sequence goes 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, and so forth. Written as a rule, the expression is xn = xn-1 + xn-2.
o Write a Java program that generates Fibonacci sequence. Your program must implement the following:
- It accepts an int value from the user.
- Starting with 0 and 1, it prints the Fibonacci sequence until the number (if the number
from the user is part of the sequence) or until the largest sequence number no bigger
than the user input (if the number from the user is NOT part of the sequence)
- A sample demo screenshot:
Enter a number: 23
0, 1, 1, 2, 3, 5, 8, 13, 21
o Review and clarify user requirements with the assessor if required
o Apply correct Java syntax rules and best practices, e.g. correct Java statements and logic o NOTE: Your code format and standards are assessed against the Java C