WordWriterFromUser
You are required to develop Java programs that demonstrate file input and output.
o The first program will accept a single word (without a space) from the user and save it in a text
file named message.txt. Save the programme as WordWriterFromUser.java
o The second program will read the word from the text file, message.txt and display it on the
screen. Save the program as WordReaderFromFile.java