EggsOrder
• WholeWorths sells eggs to local customers. They charge $7.25 for a dozen eggs, or $0.75 for an egg
not included in a dozen. This business needs an application that calculates the total price.
• During this activity, your team will design and develop an application for the business, and will prepare
and execute test procedures for the application.
• Application requirements:
o Develop a program that prompts the user for the number of eggs in the order and then displays the amount owed with a full explanation.
o Use appropriate data types and operators with correct expressions in order to create clear and concise code in your program.
o Review and clarify user requirements with the assessor if required
o Include application document in your code using appropriate block code comments o Save the program as EggsOrder.java
o A typical output might be as below if the user ordered 26 eggs.
You ordered 26 eggs. That is 2 dozen eggs at $7.25 per dozen and 2 additional eggs
at $0.75 each for a total of $16.00.